[ld] Add pass4
This pass is combining every section in a
temporary file. This is only a skeleton
because it doesn't care about alignment
or rellocations.

+++ src/cmd/ld/ld.h
@@ -8,8 +8,8 @@ typedef struct segment Segment;
 
 struct section {
 	char *name;
-	unsigned long base;
-	unsigned long long size;
+	unsigned long long base;
+	unsigned long size;
 	unsigned flags;
 	int type;
 	FILE *fp;
